[TOK-14] Fix migration file + rename auth options file
