+++ src/app/api/auth/[...nextauth]/index.ts
@@ -1,4 +1,4 @@
 import NextAuth from "next-auth/next";
-import authOptions from "@/lib/authentication/options";
+import authOptions from "@/lib/authentication/authOptions";
 
 export default NextAuth(authOptions);
