+++ src/lib/authentication/authOptions.ts
