+++ src/lib/authentication/session.ts
@@ -1,6 +1,6 @@
 import type { CustomSession } from "@/model/authentication";
 import { getServerSession } from "next-auth";
-import authOptions from "./options";
+import authOptions from "./authOptions";
 
 export const getSession = async () => {
   return (await getServerSession(authOptions)) as CustomSession;
